requirements.txt
chainlit
langchain_groq
python-dotenv
langchain_community
numexpr